# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Curt API",
        "contact": {
            "name": "Salvatore Emilio",
            "url": "http://salvatoreemilio.it",
            "email": "@info@salvatoreemilio.it"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.2.0"
    },
    "paths": {
        "/c": {
            "get": {
                "security": [
                    {
                        "X-API-Key": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "c"
                ],
                "summary": "List all Curt(s)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Curt"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GenericError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "X-API-Key": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "c"
                ],
                "summary": "Create a new Curt",
                "parameters": [
                    {
                        "description": "Curt Data",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Body"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Curt"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.GenericError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GenericError"
                        }
                    }
                }
            }
        },
        "/c/{key}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "c"
                ],
                "summary": "Follow a Curt redirect",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Curt Key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "301": {
                        "description": "Moved Permanently"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.GenericError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GenericError"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "c"
                ],
                "summary": "Delete a Curt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Curt Key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Curt"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.GenericError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GenericError"
                        }
                    }
                }
            }
        },
        "/status/about": {
            "get": {
                "security": [
                    {
                        "X-API-Key": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "status"
                ],
                "summary": "About",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Module"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GenericError"
                        }
                    }
                }
            }
        },
        "/status/health": {
            "get": {
                "security": [
                    {
                        "X-API-Key": []
                    }
                ],
                "produces": [
                    "plain/text"
                ],
                "tags": [
                    "status"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GenericError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Body": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "TTL": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "models.Curt": {
            "type": "object",
            "properties": {
                "TTL": {
                    "type": "integer"
                },
                "curt": {
                    "type": "string"
                },
                "expiresAt": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "models.GenericError": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.Module": {
            "type": "object",
            "properties": {
                "info": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "replace": {
                    "$ref": "#/definitions/models.Module"
                },
                "sum": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "X-API-Key": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}